pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "toxic-news"
description = 'Evaluating toxicity of news'
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
keywords = []
authors = [
    { name = "Daniel Ferreira", email = "daniel.ferreira.1@gmail.com" },
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
    "detoxify==0.5.1",
    "lxml==4.9.2",
    "requests==2.28.2",
    "loguru==0.6.0",
    "pydantic==1.10.6",
    "Jinja2==3.1.2",
    "pymongo==4.3.3",
    "python-dotenv==1.0.0",
    "fastapi==0.94.1",
    "uvicorn==0.21.0",
    "typer[all]==0.7.0",
    "aiohttp==3.8.4",
]
dynamic = ["version"]

[project.urls]
Documentation = "https://github.com/dcferreira/toxic-news#readme"
Issues = "https://github.com/dcferreira/toxic-news/issues"
Source = "https://github.com/dcferreira/toxic-news"

[tool.hatch.version]
path = "toxic_news/__about__.py"

[tool.hatch.envs.default]
dependencies = [
    "pytest",
    "pytest-cov",
    "pytest-snapshot",
    "pytest-xprocess",
    "black",
    "isort",
    "ruff",
    "pre-commit",
    "mypy",
    "types-requests",
    "lxml-stubs",
]
[tool.hatch.envs.default.scripts]
cov = "pytest --cov-report=term-missing --cov-config=pyproject.toml --cov=toxic_news --cov=tests {args}"
no-cov = "cov --no-cov {args}"
integration = "no-cov -m integration"
slow = "no-cov -m slow"
lint = ["ruff .", "black . --check -q", "isort . --check -q"]
format = ["black .", "isort ."]
types = "mypy ."
generate = "python toxic_news/generate.py"
build = "docker buildx build . -t toxic-news:local"
serve-docker = "docker compose up"
push = [
    "docker tag toxic-news:local us-central1-docker.pkg.dev/toxic-news/toxic-news/toxic-news:latest",
    "docker push us-central1-docker.pkg.dev/toxic-news/toxic-news/toxic-news:latest"
]

[tool.hatch.envs.explore]
dependencies = [
    "jupyterlab",
    "seaborn",
    "ipympl",
]

[tool.hatch.envs.explore.scripts]
run = "jupyter lab --no-browser"

[[tool.hatch.envs.test.matrix]]
python = ["39", "310", "311"]

[tool.coverage.run]
branch = true
parallel = true
omit = [
    "toxic_news/__about__.py",
]

[tool.coverage.report]
exclude_lines = [
    "no cov",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[tool.isort]
profile = "black"

[tool.mypy]
check_untyped_defs = true

[[tool.mypy.overrides]]
module = [
    "detoxify",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = "--strict-markers -m \"not integration\" -m \"not slow\""
markers = [
    "slow: slow and heavy tests, not sure if these can run in the CI pipeline",
    "integration: slow tests, shouldn't be run so often"
]
